# ApplicationsDashboard-Api/src/main/resources/api.yaml
swagger: '2.0'
info:
  title: Applications Dashboard
  description: Applications Dashboard
  version: "0.0.1"
basePath: /
produces:
  - application/json
paths:
  /ping:
    get:
      operationId: ping
      summary: Application HealthCheck
      tags:
        - Info
      responses:
        200:
          description: Ok Object
  /application:
    get:
      operationId: getApplications
      summary: Result a list with all Applications
      tags:
        - Application
      responses:
        200:
          description: List of applications
          schema:
            type: array
            items:
              $ref: '#/definitions/Application'
    post:
      operationId: createApplication
      summary: Create application
      tags:
        - Application
      parameters:
        - name: application
          in: body
          require: true
          schema:
            $ref: '#/definitions/ApplicationCreate'
      responses:
        201:
          description: Application Created
        405:
          description: "Invalid input"
  /application/{appId}:
    get:
      operationId: getApplication
      summary: Return Application Data
      tags:
        - Application
      parameters:
        - name: appId
          type: integer
          format: int64
          description: Application Identifier
          in: path
          required: true
      responses:
        200:
          description: Application Info
          schema:
            type: object
            $ref: '#/definitions/Application'
        404:
          description: Application Not Found
    put:
      operationId: updateApplication
      summary: Update application
      tags:
        - Application
      parameters:
        - name: appId
          type: integer
          format: int64
          description: Application Identifier
          in: path
          required: true
        - name: application
          in: body
          require: true
          schema:
            $ref: '#/definitions/ApplicationUpdate'
      responses:
        204:
          description: Application Updated
        404:
          description: "Invalid input"
  /application/{appId}/serviceInfo:
    get:
      operationId: getServiceInfo
      summary: Return Service Information
      tags:
        - ServiceInfo
      parameters:
        - name: appId
          type: integer
          format: int64
          description: Application Identifier
          in: path
          required: true
      responses:
        200:
          description: Service Info
          schema:
            type: object
            $ref: '#/definitions/ServiceInfo'

  /application/{appId}/endpointsInfo:
    get:
      operationId: getEndpointsInfo
      summary: Return Endpoints Information
      tags:
        - ServiceInfo
      parameters:
        - name: appId
          type: integer
          format: int64
          description: Application Identifier
          in: path
          required: true
      responses:
        200:
          description: Endpoints Information
          schema:
            type: array
            items:
              $ref: '#/definitions/EndpointInfo'

definitions:
  Application:
    type: object
    properties:
      id:
        type: string
        description: Application Unique Identifier
      name:
        type: string
        description: Application Name
      dns:
        type: string
        description: Application DNS (protocol must be present)
      healthEndpoint:
        type: string
        description: Endpoint used to validate application healthy
  ApplicationCreate:
    type: object
    properties:
      name:
        type: string
        description: Application Name
      dns:
        type: string
        description: Application DNS (protocol must be present)
      healthEndpoint:
        type: string
        description: Endpoint used to validate application healthy
  ApplicationUpdate:
    type: object
    properties:
      name:
        type: string
        description: Application Name
      dns:
        type: string
        description: Application DNS (protocol must be present)
      healthEndpoint:
        type: string
        description: Endpoint used to validate application healthy

  ServiceInfo:
    type: object
    properties:
      healthy:
        type: boolean
        description: Application Health Status
      version:
        type: string
        description: Service Version

  EndpointInfo:
    type: object
    properties:
      endpoint:
        type: string
        desciption: Endpoint
      result:
        type: string
        desciption: Response from endpoint